Refactor postprocessTree logic
